Set up models django admin. Also made further changes to settings.py
Populate the models like these:

Preferred Contact - Email address, Mobile

Time Choices - Upto you, follow this format: 10:00 PM

Status - Pending, Completed, Cancelled

Populating these models will allow you to access them as a dropdown in the Appointment model

Create a user first in the User model so that you can add them in the Agent and Customer models.

This goes without saying, but tick is agent and untick is organizer when adding an agent in the User model.

Untick both if you're adding a customer into the User model